requirements/base.txt
Django==2.2.13
django-filter==2.2.0
django-health-check==3.6.1
git+https://github.com/buildlyio/django-oauth-toolkit-jwt@v0.5.2#egg=django-oauth-toolkit-jwt
djangorestframework==3.9.4
psycopg2-binary==2.8.6
social-auth-app-django==3.1.0
django-oauth-toolkit==1.3.2
futures==3.1.1
django-cors-headers==2.5.3
pyswagger==0.8.39
bravado-core==5.13.1
drf-yasg==1.10.2
requests==2.21.0
aiohttp==3.5.4
django-auth-ldap==2.1.0
rest-social-auth==5.0.0